Ensure cachefile object is not None before trying to close it
Since the cachefile object can be None under certain conditions,
check for this before trying to close the cachefile handle.
This should help in clarifying what cased the error handler to be called
in the first place since the log will not have an extra exception that is unrelated
to the real exception.